Move utils to new file
